+++ dbo.spPAY_GetLoanRepayAmountDetails.sql
@@ -3,7 +3,7 @@ GO
 SET ANSI_NULLS, QUOTED_IDENTIFIER ON
 GO
 CREATE PROCEDURE [dbo].[spPAY_GetLoanRepayAmountDetails]
-	@DOCID [int],
+	@DOCID [bigint],
 	@UserID [int] = 1,
 	@LangID [int] = 1
 WITH ENCRYPTION, EXECUTE AS CALLER
