+++ dbo.spPAY_GetPayrollCutomization.sql
@@ -3,7 +3,7 @@ GO
 SET ANSI_NULLS, QUOTED_IDENTIFIER ON
 GO
 CREATE PROCEDURE [dbo].[spPAY_GetPayrollCutomization]
-	@GradeID [int],
+	@GradeID [bigint],
 	@PayrollDate [datetime],
 	@UserID [int] = 0,
 	@LangID [int] = 1
@@ -15,7 +15,5 @@ BEGIN
 	SELECT REPLACE(RIGHT(CONVERT(VARCHAR(11), Convert(DATETIME,MAX([PayrollDate])), 106), 8), ' ', '-') PayrollDate,isnull(MAX(SNO),0) SNO FROM [COM_CC50054] WITH(NOLOCK)
 	
 	SELECT * FROM PAY_PayrollPT WITH(NOLOCK) WHERE [PayrollDate]=@PayrollDate
-
-	
 END
 GO
