+++ dbo.spPAY_SetGradeWiseSalarySlabs.sql
@@ -3,7 +3,7 @@ GO
 SET ANSI_NULLS, QUOTED_IDENTIFIER ON
 GO
 CREATE PROCEDURE [dbo].[spPAY_SetGradeWiseSalarySlabs]
-	@GradeID [int] = NULL,
+	@GradeID [bigint] = NULL,
 	@WEF [datetime] = NULL,
 	@GSXML [nvarchar](max) = NULL,
 	@UserName [nvarchar](50)
@@ -24,14 +24,14 @@ BEGIN
 	DELETE FROM  PAY_GradeWiseSalarySlabs WHERE GradeID=@GradeID and WEF=@WEF
 	
 	INSERT INTO PAY_GradeWiseSalarySlabs(GradeID,WEF,SNo,ComponentID,MinAmount,MaxAmount,AlertType,ApplicableFormula,CreatedBy,CreatedDate,ModifiedBy,ModifiedDate)
-	SELECT @GradeID,@WEF,CONVERT(NVARCHAR,A.value('@SNo','INT')),CONVERT(NVARCHAR,A.value('@ComponentID','INT')),CONVERT(float,A.value('@MinAmount','float')),CONVERT(float,A.value('@MaxAmount','float')),CONVERT(NVARCHAR(MAX),A.value('@AlertType','nvarchar(max)')),CONVERT(NVARCHAR(MAX),A.value('@ApplicableFormula','nvarchar(max)')),@UserName,@SysDate,@UserName,@SysDate
+	SELECT @GradeID,@WEF,CONVERT(NVARCHAR,A.value('@SNo','INT')),CONVERT(NVARCHAR,A.value('@ComponentID','BIGINT')),CONVERT(float,A.value('@MinAmount','float')),CONVERT(float,A.value('@MaxAmount','float')),CONVERT(NVARCHAR(MAX),A.value('@AlertType','nvarchar(max)')),CONVERT(NVARCHAR(MAX),A.value('@ApplicableFormula','nvarchar(max)')),@UserName,@SysDate,@UserName,@SysDate
 	FROM @XML.nodes('/Rows/Row') as Data(A)
 	
 END
 ELSE
 BEGIN
 	INSERT INTO PAY_GradeWiseSalarySlabs(GradeID,WEF,SNo,ComponentID,MinAmount,MaxAmount,AlertType,ApplicableFormula,CreatedBy,CreatedDate,ModifiedBy,ModifiedDate)
-	SELECT @GradeID,@WEF,CONVERT(NVARCHAR,A.value('@SNo','INT')),CONVERT(NVARCHAR,A.value('@ComponentID','INT')),CONVERT(float,A.value('@MinAmount','float')),CONVERT(float,A.value('@MaxAmount','float')),CONVERT(NVARCHAR(MAX),A.value('@AlertType','nvarchar(max)')),CONVERT(NVARCHAR(MAX),A.value('@ApplicableFormula','nvarchar(max)')),@UserName,@SysDate,@UserName,@SysDate
+	SELECT @GradeID,@WEF,CONVERT(NVARCHAR,A.value('@SNo','INT')),CONVERT(NVARCHAR,A.value('@ComponentID','BIGINT')),CONVERT(float,A.value('@MinAmount','float')),CONVERT(float,A.value('@MaxAmount','float')),CONVERT(NVARCHAR(MAX),A.value('@AlertType','nvarchar(max)')),CONVERT(NVARCHAR(MAX),A.value('@ApplicableFormula','nvarchar(max)')),@UserName,@SysDate,@UserName,@SysDate
 	FROM @XML.nodes('/Rows/Row') as Data(A)
 END
 
@@ -47,4 +47,5 @@ Begin Catch
 	SELECT ERROR_MESSAGE() AS ErrorMessage
 	ROLLBACK transaction
 End Catch
+
 GO
